Change PluginBase._state_items from list to dict
The PluginBase object stored state_items as a list of (path, getter,
setter) tuples.  The StatusManager plugin collected these into one
place by passing a 'result' list to each plugin, which would append
its own respective list of tuples, a rather unusual way of returning
data back to a calling function.  The plugins used 'if ... in ...' and
'list.remove()' on the lists, also unusual.

This commit changes state items to a dict keyed on path, with the
value a (getter, setter) tuple.  Locating and clearing state items by
path is more intuitive with a dict.  In addition, the resulting dict
can be serialized directly by e.g. PyYAML to create a human-readable,
structured configuration file or other configuration persistence.