Settings persistence:  remove _state_item & get settings from self.core
The redundant _state_item infrastructure has been removed.

Instead, non-ListPluginBase plugins that have settings to be saved
should mark them in either the 'PERSIST_GENERAL_PREFERENCES' or
'PERSIST_TASK_SETTINGS' attribute.  These should be in a dict to
structure the data (dict of dicts to structure it deeply), and values
should be lists of self.core keys to persist.

StatusManager has a new method get_global_persist_data.  This takes
one argument: the name of the PERSIST_* attribute noted above.  It
goes through each plugin in turn, calling its get_persist_data method
with the same argument, and merging the results.  This data is
guaranteed to be only built-in Python structures and data types to
make it easier for a variety of picklers to deal with.

For now, colors are marked for persistence here, but this should be
changed soon.

Next steps:
     - plug this into saving
     - fix up restoring and plug this in there