Add a command-line debugging console
A menu entry in the 'View' menu (for lack of a better place) starts a
python interactive shell on the console.  This is very handy for newby
developers trying to find their way around (I know this personally!).

FIXME:

It will break the GtkConsole until a proper solution for stdin/stdout
handling is devised.  The GtkConsole hijacks those, which breaks the
console shell.

FIXME:

The menu item needs a weight to appear at the bottom of the list.
Right now it appears randomly at top or bottom.

FIXME:

Alternatively, add some extra buttons to the existing graphical
console to add debugging objects (self.core) and start the shell on
the terminal.