Plugins/__init__.py:  add warning about potential pygtk or gtk bug
treemodel.reorder() seems to be broken in fc16 with pygtk 2.24.0 and
gtk 2.24.8.  Experimentation showed that the order of items in the
ListStore object didn't change after a reorder() with a valid
sorted_indices list.

This patch prints a warning if the treemodel is not sorted properly
after the reorder().

FIXME